Charttest mod

+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/vo/StorageFreeVo.java
@@ -10,7 +10,7 @@ public class StorageFreeVo {
     private Date date;
     private String storage_name;
     private BigInteger daily_used;
-    private BigInteger free;
+    private float free;
     private BigInteger total;
     private String start_date;
     private String end_date;
